RaspberryPi4/requirements.txt
#camerastream
picamera2
fastapi
uvicorn
opencv-python

#servo
paho-mqtt
adafruit-circuitpython-servokit
